changes by frontend

+++ src/main/java/sfa/order_service/dto/response/OutletReportResponse.java
@@ -11,5 +11,5 @@ import lombok.experimental.FieldDefaults;
 public class OutletReportResponse {
     Double totalSales;
     OutletRespForOrderDto outletRespForOrderDto;
-    Integer totalOrder;
+    Long totalOrder;
 }
